Fix broken notification emails
After support for showing the context around a notification message, the
code used LOAD-MESSAGE-RANGE-WITH-CHECK to load the message range. The
problem is that this function needs a current user context in order to
be able to check is the user has access to the channel.

There is no need to check that the user is in the channel, since the
notification object is only ever created for users that are in the
channel in the first place. Therefore, a new function LOAD-MESSAGE-RANGE
was created that does not perform that access check, and therefore does
not need a current user context. The email generation code was changed
to use this new function.