Remove requirement on ES6
In the function encode-string-as-sha1, the character codes were
retrieved using .codePointAt() which is only available in ES6.
However, in this particular function there is no need to use this
function since the characters are always limited to 8-bit entities
anyway. This change replaces this call with a call to .charCodeAt()
which should be available everywhere.