Make image conversion more stable
In some cases the output of the convert program may contain extra rows
of data. This confused the code that parsed this output in order to
determine the final dimensions of the image.

The converter now looks at all output in order to find the correct
line to parse, instead of only checking two specific lines.